+++ packages/language-service/src/messages/warn.ts
@@ -7,3 +7,10 @@ export const GlobalTypeIsNonObjectJs = (name: string): LSMessage => {
         "https://qingkuai.dev/misc/typescript.html#component-attribute-types"
     ]
 }
+
+export const KeyframeWillNotBeScoped = (): LSMessage => {
+    return [
+        7002,
+        "The @keyframes rule is not scoped. It is recommended to define it in an external stylesheet and import it from the application entry file."
+    ]
+}